% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srd_functions.R
\name{srd.plot}
\alias{srd.plot}
\title{To  plot an srd drawing object}
\usage{
srd.plot(fit)
}
\arguments{
\item{fit}{An output object of  \code{srd}}
}
\description{
To  plot an srd drawing object
}
\details{
A function to plot an  srd diagram after creating with srd() 
 \code{\link{srd}}.
}
\author{
Roger Marshall <rj.marshall@auckland.ac.nz>, The University of Auckland, New Zealand
}
