% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srd.r
\name{srd}
\alias{srd}
\title{Draws a scaled rectangle diagram.}
\usage{
srd(
  data,
  rectangles = NULL,
  weight = NULL,
  new = TRUE,
  seed = NA,
  whitespace = TRUE,
  criterion = "labs",
  thin = 0,
  z = NULL,
  title = NULL,
  layer = TRUE,
  clickable = TRUE,
  reweight = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame or a table}

\item{rectangles}{A character vector to represent the \eqn{k} rectangles. 
Each element is a valid logical expression that defines a rectangle (see examples).}

\item{weight}{A string giving name of a frequency weight of each row of \code{data}.}

\item{new}{If \code{TRUE}  a new configuration is generated. \code{new=FALSE}  assumes 
a configuration from a previous matching \code{srd()} call already exists in the workspace.}

\item{seed}{Random number seed. \code{NA}  for non-repeatable, possibly different configurations.}

\item{whitespace}{If FALSE will not display whitespace, that is, 
the space that is not in at least one of the rectangles specified in \code{rectangles}.}

\item{criterion}{Criterion to determine fitting optimisation criterion.
Use \code{"lsqs"} for least squares, \code{"labs"} for least absolute difference, \code{"logl"} for minus log-likelihood, 
\code{"chisq"} for chi-square, \code{"stress"} for constrained least squares.}

\item{thin}{A penalty parameter to avoid thin and elongated rectangles. The \code{criterion} is penalized 
by \eqn{e^thin} where \eqn{e} maximal length/breadth of the  rectangles.}

\item{z}{The name of variable(s) to form a 3D projection axis. If a single variable
the mean value of the variable is projected, if  two variables the ratio of their means
is projected. If \code{NA}, the sequence number of rectangles is assigned.}

\item{title}{A title written on the srd graphic.}

\item{layer}{Determines layering of rectangles from back to front. If TRUE smallest rectangle is at front
largest at back. If FALSE, layered as the order listed in \code{rectangles} parameter with first at back last at front.
layer is ignored if \code{z} is non-NULL, when layering is by value of \code{z} (largest on top).}

\item{clickable}{\code{TRUE} if the graphic is active for on-the-fly mouse input (see Details).}

\item{reweight}{a character string that matches one of \code{rectangles} 
to force analysis of pseudo counts in which the \code{reweight} rectangle
is  forced relative frequency 0.5.}

\item{...}{Other control parameters, for example user defined
rectangle colours (see details).}
}
\value{
A summary of the graphic.  It gives rectangle and cell 
statistics and computational notes. Use
\code{\link{srd.summary}} to display.
}
\description{
Draws a scaled rectangle diagram.
}
\details{
A function to visualise  \eqn{k} attributes as a Venn-like  
diagram using \eqn{k}  rectangles
with areas scaled, as best as possible, to be proportional to  frequency. 
It allows \eqn{k \le 6}.  
Fitting is done by optimising the congruence between cell 
area \eqn{a_i} and cell frequency  \eqn{ n_i} 
by a chosen  criterion.
A starting configuration is fixed by intelligent or
 random positioning. A graphic is produced that is active for mouse 
 input if \code{clickable=TRUE} with a temporary diaglog box.
 Dialog options include: \code{Refit} which toggles fitting criterion to 
 usually created a different configuration;  \code{colour} which toggles 
 different colour shading arrangements, one of which is random colours (for user specified 
 colours see \code{...} parameters);  \code{Mode} which toggles "cell" 
 and "rectangle" mode (the distinction is that, 
 in an oblique projection, either cells or rectangles are 
 rendered); \code{Cell stats} superimposes statistics of 
 cells; \code{Labels} toggles how labels to rectangles are drawn;  \code{Rotate} 
  throws the diagram into  a 3D projection allowing rotation steps,
 the third axis being \code{z}.  
 Repeated clicks turns the diagram. This is more quickly 
 done with \code{Toggle view} which toggles an oblique, 
 flat or full side-on view.  \code{Esc}  leaves the mouse-active graphic. 
 
Additional ellipsis \code{...} parameters may include 
 a colour pallette for rectangles  \code{col}. 
Each  colour may include a transparency code, otherwise rectangles are opaque.
If a non-white colour for whitespace is preferred it can be specified with \code{col0}.
Background colour is specified with \code{bg}.
}
\examples{
# Simulate  and show independent binary A, B, C 
A <- rbinom(1000,1,0.1)
B <- rbinom(1000,1,0.4)
C <- rbinom(1000,1,0.3)
x <- cbind(A,B,C)
srd(data=x,rectangles=c("A","B","C"))

# Titanic data
library(datasets)
 Titan <- srd(data=Titanic,
     rectangles=c("Sex=='Female'",
                  "Age=='Child'",
                  "Class=='1st'"),
       weight="Freq", title="Titanic survival",z="Survived")
srd.summary(Titan) 

## Co-occurrence of Scabies  and Impetigo in Fiji 
scab_imp <- data.frame( "Scabies"=c(1,1,0,0),"Impetigo"=c(1,0,1,0),"n"=c(2021,543,112,8211))
srd(rectangles=c("Scabies","Impetigo"),weight="n",
           clickable=TRUE,scab_imp, criterion="chisq",
           col=c("#ff0012b3","#765388b3") )
 
}
\references{
Marshall, RJ. Displaying  categorical data relationships by scaled rectangle diagrams. Statistics in Medicine 20, 1077-1088, 2001.
Marshall, RJ. Scaled rectangle diagrams can be used to visualise clinical and epidemiological data. 
Journal of Clinical Epidemiology 58, 974-981, 2005.
Marshall, RJ. Determining and visualising at-risk groups in case-control data. 
Journal of Epidemiology and Biostatistics 6,  343-348, 2001
}
\seealso{
\code{\link{srd.summary}}  for summary statistics of an \code{srd} object
}
\author{
Roger Marshall, <rj.marshall@auckland.ac.nz>, The University of Auckland, New Zealand
}
