% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srd_functions.R
\name{srd.summary}
\alias{srd.summary}
\title{To  summarise a n srd object}
\usage{
srd.summary(fit)
}
\arguments{
\item{fit}{An output object of  \code{srd}}
}
\description{
To  summarise a n srd object
}
\details{
A function to summarise the result of a run of 
 \code{\link{srd}}.
}
\author{
Roger Marshall <rj.marshall@auckland.ac.nz>, The University of Auckland, New Zealand
}
